requirements.txt
allure-pytest==2.13.5
allure-python-commons==2.13.5
attrs==24.2.0
black==24.8.0
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
data==0.4
decorator==5.1.1
funcsigs==1.0.2
idna==3.7
iniconfig==2.0.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pluggy==1.5.0
pytest==8.3.2
requests==2.32.3
six==1.16.0
urllib3==2.2.2
